Close the QA locked lot: authority is minted where it is earned
Every posture the two audits proved wearable-without-earning now
refuses at its own constructor:

- Membership refuses a member whose role is outside R::ALL, through
  both construction roads; the closure's whole-set claim is structural.
- StampedPlan's fields are private; stamp::planned is the only mint,
  and a compile-refusal fixture proves the literal cannot be written.
- A publication address on a seat that never publishes refuses at
  planning instead of riding the identities inert.
- The transcript row law refuses delivery-before-send.
- Loom's Builder has all ten seats forced explicitly after
  construction; a valid LOOM_* environment can no longer end
  exploration early and mint the zero-run exhaustive pass.

The rest of the lot:

- Preemption is opt-in; the wall runs --all-features; the harness
  default build compiles on wasm32 again.
- Semantic keys and seat nodes carry the kind's name as a framed
  ancestor, so two kinds over one capture cannot share a unit.
- Profile, NamespacedName, and the framing primitives each have one
  encoding seat; four identifier alphabets collapsed into one
  rendered_identifier with the rust_keyword roster beside it.
- The mutation family spells the bank's declared-order-permutation.
- Direct-grammar numbers parse at their seats' exact widths; rosters,
  separators, keyword names, and generated-name collisions refuse at
  their own tokens.
- OutputAndDigest answers with the complete roster-ordered set;
  WholeDeclaration is a site without a manufactured token; a
  single-cause refusal's related set is empty.
- Both stale trybuild fixtures prove exactly their named boundaries;
  the replay-posture non-vacuity lane closes the surviving mutants'
  evidence gap; the doc pass narrows every overclaim it reached.

Wall: nextest 194/194, 0 skipped, under --all-features; check and
clippy clean across all feature postures; fmt, doc -D warnings, and
deny clean.

+++ macros/compiler/src/descriptor/concurrency/render.rs
@@ -180,9 +180,9 @@ fn explored_let(
     arguments.push(GeneratedToken::group(
         GeneratedDelimiter::Parenthesis,
         vec![
-            GeneratedToken::number(row.interleavings()),
+            GeneratedToken::number(u64::from(row.interleavings())),
             GeneratedToken::alone(','),
-            GeneratedToken::number(row.samples()),
+            GeneratedToken::number(u64::from(row.samples())),
         ],
     )?);
     arguments.push(GeneratedToken::alone('?'));
